requirements.txt
altgraph==0.17.4
bottle==0.13.2
macholib==1.16.3
packaging==24.2
proxy_tools==0.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pywebview==5.4
setuptools==75.8.0
typing_extensions==4.12.2
